pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src", "."]
include = ["backend", "miner", "validator", "core", "evaluator"]

[tool.setuptools.package-data]
"kinitro.evaluator.rpc" = ["*.capnp"]

[project]
name = "kinitro"
dynamic = ["version"]
description = "Storb RL Subnet - Shared Core Package"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.10.0",
    "asyncpg>=0.30.0",
    "colorama>=0.4.6",
    "duckdb>=1.3.2",
    "dynaconf>=3.2.11",
    "fastapi>=0.104.0",
    "fiber @ git+https://github.com/threetau/fiber.git@488277ddfcd80650e238f96f41ca38ec76c600d9#egg=fiber[chain]",
    "greenlet>=3.2.4",
    "gymnasium==1.2.0",
    "huggingface-hub>=0.34.4",
    "kubernetes>=33.1.0",
    "metaworld==3.0.0",
    "mujoco>=3.2.0",
    "numpy>=1.24.0",
    "pathlib>=1.0.1",
    "pgqueuer>=0.24.2",
    "psycopg2>=2.9.10",
    "pycapnp>=2.0.0",
    "pydantic>=2.0.0",
    "pytest>=8.4.1",
    "ray>=2.48.0",
    "rich>=13.0.0",
    "snowflake-id==1.0.2",
    "sqlalchemy>=2.0.43",
    "testing-postgresql>=1.3.0",
    "torch>=2.8.0",
    "typer>=0.9.0",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
]

[tool.setuptools.dynamic]
version = { attr = "core.__version__" }

[dependency-groups]
dev = ["debugpy>=1.8.9", "py-spy>=0.4.0", "pytest>=8.3.4", "ruff>=0.8.2"]
